requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
typing-extensions>=4.0.0

# SharePoint integration
Office365-REST-Python-Client>=2.4.0

# Image processing
Pillow>=9.0.0

# OpenAI integration
openai>=1.0.0

# Utility libraries
tqdm>=4.66.0
pytest>=7.0.0
pytest-mock>=3.10.0
mypy>=1.0.0
requests>=2.28.0
